No support for modifications within same migration
Drop support for an operation in a migration script to modify a resource
that was created earlier in the same migration.

This is not useful in practice and supporting it may lead to problems
around having later ops in a migration be aware of temporary names when
rolling back migrations.

* Change the example file to add a column to a table created by a an
  earlier migration.
* Remove the tests that check for this behaviour.