Make index name mandatory on create index operation (#59)
Make specifying a name mandatory on the **Create index** operation. 

In order to work with indexes in subsequent migrations (eg deleting the
index), the user will have to know the name of the migration. If the
migration name is auto-generated and we ever change how names are
generated, then we risk breaking a user's migrations if they have
migrations that refer to these generated names.