Implement adding uniqueness constraints to columns (#53)
Add support for adding uniqueness constraints to columns. Such a
migration looks like this:

```json
{
  "name": "15_set_columns_unique",
  "operations": [
    {
      "set_unique": {
        "name": "reviews_username_product_unique",
        "table": "reviews",
        "columns": [
          "username",
          "product"
        ]
      }
    }
  ]
}
```
This migration adds a unique constraint spanning the `username` and
`product` columns in the `reviews` table.

* On `Start` a unique index is created concurrently.
* On `Rollback` the unique index is removed.
* On `Complete` a unique constraint is added to the column using the
index.

Creating a unique constraint directly requires a full table exclusive
lock. By first creating a unique index concurrently and then adding a
constraint using the index the need for the lock is avoided.