Generalise the `createTrigger` function
Allow the trigger to take:
* an additional expression to test before applying the `up` or `down`
  SQL.
* An optional `else` clause

The extra test expresssion is useful when deciding whether to apply the
`up` SQL when writing to a column that should have `NOT NULL`
constraint. In this case the `up` SQL should be used only if the value
to be written is `NULL`.

The `else` clause is useful in the same situation when the value to be
written is not `NULL`; in this case the value should be written to the
new column unchanged.

The generated SQL looks like:

```
IF search_path <> latest_schema AND NEW."review" IS NULL THEN
  NEW."_pgroll_new_review" = product || ' is good';
ELSE
  NEW."_pgroll_new_review" = NEW."review";
END IF
```