Support column aliasing in trigger template code (#75)
The `.Columns` field provided to the template for execution has type
`map[string]schema.Column`. The existing template code ignored the key
values using just the `.Name` field in the `schema.Column`. By using the
key string, the template is able to support aliasing columns in
declarations, for example:

```go
Columns: map[string]schema.Column{
  "id":       {Name: "id", Type: "int"},
  "username": {Name: "username", Type: "text"},
  "product":  {Name: "product", Type: "text"},
  "review":   {Name: "review", Type: "text"},
  "rating":   {Name: "_pgroll_new_rating", Type: "integer"},
},
```
aliases the `rating` key to the `_pgroll_new_rating` column, producing a
SQL declaration in the trigger like:

```sql
      "rating" "public"."reviews"."_pgroll_new_rating"%TYPE := NEW."_pgroll_new_rating";
```

This is useful when generating declarations for the **change column
type** operation, where the `rating` variable in the down SQL should
refer to the new temporary column.