Move the `set_unique` operation under `alter_column` (#116)
Move the `set_unique` operation into the `alter_column` operation and
restrict it to creating constraints on single columns.

`set_unique` operations used to look like this:

```json
{
  "name": "15_set_columns_unique",
  "operations": [
    {
      "set_unique": {
        "name": "reviews_username_product_unique",
        "table": "reviews",
        "columns": [
          "username",
          "product"
        ]
      }
    }
  ]
}
```
where the operation was called `set_unique` and multi-column constraints
could be defined. The operation now looks like:


```json
{
  "name": "15_set_column_unique",
  "operations": [
    {
      "alter_column": {
        "table": "reviews",
        "column": "review",
        "unique": {
          "name": "reviews_review_unique"
        }
      }
    }
  ]
}
```
where the operation is now `alter_column` and only a single column is
covered by the constraint.

Restricting the constraint to a single column makes it possible to have
the operation duplicate the affected column and add up/down triggers
when starting a migration as we do for other `alter_column` operations.