Add a 'set replica identity' operation  (#201)
Add support for a new **set replica identity** operation type. The new
operation looks like this:

```json
{
  "name": "29_set_replica_identity",
  "operations": [
    {
      "set_replica_identity": {
        "table": "fruits",
        "identity": {
          "type": "full"
        }
      }
    }
  ]
}
```

This sets the replica identity for the `fruits` table to `FULL`. The
other supported operation types are 'nothing', 'default', and 'index'.
If the replica identity is being set to an index, the operation looks
like:

```json
{
  "name": "29_set_replica_identity",
  "operations": [
    {
      "set_replica_identity": {
        "table": "fruits",
        "identity": {
          "type": "index",
          "index": "some_index_name"
        }
      }
    }
  ]
}
```

The replica identity is set directly on the underlying table on
operation start. This means that both versions of the table exposed in
the new and old versioned views will have the new replica identity set.

The docs have been updated with the details of the new operation
[here](https://github.com/xataio/pgroll/blob/set-replica-identity/docs/README.md#set-replica-identity).