Allow to disable version schemas (#233)
There are situatiosn were creating schema verisons is not required. This
internal flag will allow to disable creating & managing these schemas
when needed.

Note: I opted to not expose this flag to the CLI as I believe it would
make create more harm & confusion than not having the feature there.