Strengthen assertions for tests that check preservation of `UNIQUE` constraints (#277)
Strengthen the tests that check for preservation of `UNIQUE` constraints
when a column is duplicated so that they can check for the failure case
in #273.

It's not enough to test that the column does not accept duplicate values
after the migration completes. Both a unique index and a unique
constraint will have that effect but we need to ensure that the
**constraint** is present on the table once the migration completes.

Duplicating a `UNIQUE` constraint is a two-step process, first creating
a `UNIQUE` index on migration start and then upgrading the index to a
constraint on migration completion. #273 occurs when this upgrade fails
and the column is left with the unique index but not the constraint.
Subsequent migrations will then fail to duplicate the non-existent
unique constraint.

Part of #273