Limit `raw_migration` to some kinds of `sql_drop`
`ddl_command_end` event triggers also run on `sql_drop` and in most cases
contain enough information for the `pgroll.migrations` table to be
populated. For example, when a column is dropped from a table, the
schema name is available via `pg_catalog.pg_event_trigger_ddl_commands`.

Only run `raw_migration` on `sql_drop` for those kinds of drops that do
not populate the schema name on `ddl_command_end` to ensure that
duplicate entries in the `pgroll.migations` table are not created.