Improve test failure message from assertion
If the column had no comment the error message if this assertion failed
would be a scan error attempting to scan a NULL into a string.

Use a pointer to a string so that columns that fail the assertion due to
having a NULL comment have a meaningful failure message.