Restrict state search path to the state schema
Restrict the search path of the connection used by the state package to
only the state schema (the schema that containts the `migrations` table).

This ensures that any column types that might reside in other schema are
always qualified with the schema name in the internal schema
representation. Without the restriction of the search path, these type
names would be unqualified if they lived in a schema that was on the
search path used by the state connection.