Chore: Add migration start message with schema name (#384)
Hi, I noticed an issue that seemed like a quick fix, so I went ahead and
made the changes.
The related issue is here: 
- https://github.com/xataio/pgroll/issues/368

In this PR, I made changes to prioritize using the `name` property in
migration file. If `name` is empty, we fall back to using the filename.

I'd appreciate it if you could take a look when you have some time.