Build static HTML page using the latest benchmark results (#467)
This uses a GitHub action to download benchmark data from S3. It then
uses a static generator to create charts per benchmark / Postgres
version. These results are then bundled up using Jekyll which publishes
the README, available at:

https://xataio.github.io/pgroll/

The benchmarks themselves can be see at
https://xataio.github.io/pgroll/benchmarks.html (and are also linked
from the README)

In order to make this work and also not require users of pgroll as a
library to import the charting library we needed to create a new module
in `/dev`. We needed to duplicate the definition of the benchmark
results structs since we couldn't import them from the main `pgroll`
module without actually publishing a new version of the module. Once
this module has been published it will be possible.

Part of https://github.com/xataio/pgroll/issues/408