Add primary key details to virtual schema
Make `OpCreateTable` add the details of the new table's primary key to
the virtual schema.

This allows subsequent `Start` or `Validate` methods of operations in
the same migration to reference the table's primary key.

For example, this change allows the `OpAddColumn.Validate` operation to
know that if the new table has a primary key that can be used for a
backfill.