Handle unconvertable `ADD CONSTRAINT .. UNIQUE`s
We need to be conservative when converting SQL statements to `pgroll`
operations to ensure that information present in the SQL is not lost
during the conversion.

There are several options possible as part of `ADD CONSTRAINT ...
UNIQUE` statements that aren't representable by the `OpCreateConstraint`
operation; in these cases we must resort to converting to an `OpRawSQL`.