Use `omissis/go-jsonschema` image for type gen
Generate types from the JSON schema using the `omissis/go-jsonschema`
image instead of the `surjection/go-jsonschema` image.

We used the `surjection/go-jsonschema` image because we needed a
feature that was not yet released in the `omissis/go-jsonschema` image:

https://github.com/omissis/go-jsonschema/pull/220

Now that the feature is available in the `omissis/go-jsonschema` `0.17`
release we can switch back to using it.