Use `omissis/go-jsonschema` image for type generation (#545)
Generate types from the JSON schema using the
official`omissis/go-jsonschema` image instead of the custom
`surjection/go-jsonschema` image.

We built and used the `surjection/go-jsonschema` image because we needed
a feature that was not yet released in `omissis/go-jsonschema`:

https://github.com/omissis/go-jsonschema/pull/220

Now that the feature is available in the `omissis/go-jsonschema` `0.17`
[release](https://github.com/omissis/go-jsonschema/releases/tag/v0.17.0)
we can switch back to using it.