Handle `DEFERRABLE` and `NOT DEFERRABLE` constraints in `CREATE TABLE` statements (#556)
Handle `DEFERRABLE` and `NOT DEFERRABLE` constraints in `CREATE TABLE`
statements

`NOT DEFERRABLE` constraints are the default, so add testcases to ensure
that constraints with this modifier are converted to `OpCreateTable`
operations.

`DEFERRABLE` constraints and the `INITIALLY IMMEDIATE` and `INITIALLY
DEFERRED` modifiers are not supported, so fall back to raw SQL
operations when these options are present.