Reduce duplication between `CREATE TABLE` and `ADD COLUMN` SQL conversion (#562)
There is a lot of duplication between the `CREATE TABLE` and `ALTER
TABLE ADD COLUMN` conversion code.

Reduce the duplication by having the `ALTER TABLE ADD COLUMN` conversion
code and the `CREATE TABLE` conversion code use the same
`convertColumnDef` function, which handles the conversion of a column
definition to a `migrations.Column` struct along with the conversion of
the column constraints.