Use pointers for `schema`  and `Table` struct fields (#574)
Update the `schema` and `Tables` structs in the `schema` package to use
pointer types in fields.

As part of #239, the in-memory schema representation will need to be
updated more often and more accurately by individual operations in a
migration. Using pointers in these `schema` package structs makes this
possible.

Part of https://github.com/xataio/pgroll/issues/239