Soft delete tables in the virtual schema
Add a `Deleted` field to the `Table` struct in order to
allow `OpDeleteTable` to soft delete tables in the virtual schema.

Update `RemoveTable` to set the `Deleted` field to `true` instead of
removing a table from the `Tables` map.

Add an `UnRemoveTable` method to allow tables to be unremoved. This can
be used by `OpDeleteTable` to rollback the effects of a table deletion
in the virtual schema.

When looking up a table in the schema, ignore tables that have been
marked as deleted.

When creating views in the new schema, skip tables that have been marked
as deleted.