Soft delete columns in the virtual schema
Add a `Deleted` field to the `Column` struct in order to allow
`OpDropColumn` to soft delete columns in the virtual schema.

Update `RemoveColumn` to set the `Deleted` field to `true` instead of
removing a column from the `columns` map.

Add an `UnRemoveColumn` method to allow columns to be unremoved. This can
be used by `OpDropColumn` to rollback the effects of a column deletion
in the virtual schema.

When looking up a column in the schema, ignore columns that have been
marked as deleted.

When creating views in the new schema, skip columns that have been marked
as deleted.