Rollback migration operations in reverse order (#586)
Ensure that operations within a migration are rolled back in reverse
order when rolling a back a multi-operation migration.

To make this possible, rollbacks now work like this:

On migration rollback:
* Obtain the previous version of the schema (eg, if we are rolling back
`02_bar`, read the schema that resulted from the application of
`01_foo`). This uses the new `state.SchemaAfterMigration` method.
* Replay the migration `02_bar` on this schema, using a fake DB
connection to build up the changes to the virtual schema `s` without
altering the physical schema. This ensures that any indirections and
temporary columns are present in the virtual schema `s`.
* Invoke `Rollback` on each operation in reverse order, passing `s` as
the schema argument. Each `Rollback` may then update `s` to undo the
changes to the virtual schema made by the operation, ensuring that
preceding operations in the migration see the virtual schema as it was
when the operation was applied.

Part of https://github.com/xataio/pgroll/issues/239