Multi-operation migration support for `drop_table` operations (#587)
Ensure that `drop_table` operations can be used as part of
multi-operation migrations:

Add testcases for:
* **Create table, rename table, drop table**: 
* Make `OpRenameTable.Validate` update the in-memory schema so that the
following operation can validate using the new name of the table
* Make`OpDropTable` soft delete the table from the virtual schema, so
that `OpDropTable.Rollback` can un-drop the table, making it visible to
preceding operations in the table when the migration is rolled back.
* Make `OpRenameTable` undo the table rename, so that the table is
visible under its old name for preceding operations in the rollback.
* **Drop table, rename table** and **drop table, drop table**:
* Make `OpDropTable.Validate` remove the table from the virtual schema
so that validation of the subsequent operation fails.

Theses changes ensure that `drop_table` operations can be used as part
of multi-operation migrations.

Part of https://github.com/xataio/pgroll/issues/239